v2.0 release - THE DND UPDATE 6/6/20
This was a MASSIVE update, over doubling the lines of code from an already big file of 160 lines to a whopping ~340 lines, and that doesn't even include the separate file that contains a bunch of lines of text for the help menu and other smaller stuff like that. Not only are there more commands that do more, but I've changed the way every command is coded, so less messageless errors will occur, and the help menu is a billion times more helpful. Also, this'll be the first version that I'm gonna be running indefinitely off my laptop. This'll probably cause issues, but I'm hoping everything runs smoothly. I hope you guys enjoy, and make sure to send me suggestions on what else the bot should be able to do!

New Commands:
	!init
		*An initiative tracker! This is easily the most complex command written so far, and probably the most useful
		*Make sure to use ~help init since I made sure to explain as much as I could, since its so complex
	!condition
		*In addition to the initiative tracker, there's now a condition tracker to go along with it!
		*This tracker works in conjunction with the initiative tracker, so it won't work if no initiative is being tracked
	!role
		*a combination of the previous ~roles and ~reqrole commands (now !role view and !role request respectively)
		*now also includes a removal command - !role remove
	add a secret command only I can use ;)

Command edits:
	!roll
		*now takes the die size with or without the d ;)
		*now also takes modifiers, whether negative or positive
		*also also, it show exactly the roll being performed
	!help is a lot more robust, especially for commands with multiple subcommands
	add a bunch of more in depth help text to a bunch of commands

Misc:
	changed command prefix from ~ to !
	now running indefinitely off a separate "server"
	move commands into cogs and groups
		*front-end-wise, you guys will probably see no big difference, but this is pretty cool back-end-wise
		*biggest change users will see is through a much better help menu
	made a separate module file to declutter the main rotombot file of random text
		*literally no one but me is gonna care about this LOL
	edit member join message to accomodate role function edits
	edit on_ready message formatting